rename: db 저장 로직 클래스, 메서드 이름 변경
