+++ src/main/java/com/codeit/findex/scheduler/IndexApiScheduler.java
@@ -1,6 +1,6 @@
 package com.codeit.findex.scheduler;
 
-import com.codeit.findex.data.ApiAutoSyncService;
+import com.codeit.findex.data.AutoIndexDataSyncService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -9,13 +9,13 @@ import java.time.LocalDateTime;
 
 @Component
 @RequiredArgsConstructor
-public class ApiScheduler {
-    private final ApiAutoSyncService apiAutoSyncService;
+public class IndexApiScheduler {
+    private final AutoIndexDataSyncService autoIndexSyncService;
 
 //    @Scheduled(initialDelay = 5000) // 테스트용 어노테이션
     @Scheduled(cron = "0 0 0 * * *")
     public void task() throws Exception {
         System.out.println("오전 9시 30분 지수 데이터 연동:" + LocalDateTime.now());
-        apiAutoSyncService.fetchAndProcessAll();
+        autoIndexSyncService.fetchAndProcessAll();
     }
 }
